desktop/flask/requirements.txt
click==8.0.1
dnspython==1.16.0
Flask==2.0.1
Flask-PyMongo==2.3.0
Flask-WTF==0.15.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pymongo==3.12.0
Werkzeug==2.0.1
WTForms==2.3.3
SQLAlchemy==1.4.48
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
psycopg2-binary==2.9.7
